utils/requirements.txt
google-cloud-storage>=2.10.0
setuptools>=69.0.0
wheel>=0.42.0
pathlib>=1.0.1